+++ ChangeLog.md
@@ -15,6 +15,7 @@
     - Obsoletize Modal.beforeClose() replaced with Modal.set() (todo #30)
     - Obsoletize Modal.title() replaced with Modal.set() (todo #33)
     - Obsoletize Modal.setFooter() replaced with Modal.set() (todo #35)
+    - Obsoletize Modal.setClasses() replaced with Modal.set() (todo #32)
 
 ### 1.6.0